requirements.txt
# Core dependencies for Red Teaming Pipeline Universal

# PDF processing
pypdf>=3.17.0

# OpenAI client for OpenRouter API
openai>=1.12.0

# HuggingFace datasets
datasets>=2.16.0

# Environment variable management
python-dotenv>=1.0.0

# Async support (included in Python 3.7+, but listing for clarity)
# asyncio is part of standard library





art
